[#9] feat: IssueController 정리
이슈 열린, 닫힌 상태에 따른 메서드도 추가함
